Deprecate old timeline widgets
