+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/_deprecated/PostsByTimeWidget/PostsByTimeCard.tsx
@@ -3,23 +3,23 @@ import React from 'react';
 import { Box } from '@citizenlab/cl2-component-library';
 import moment from 'moment';
 
-import Chart from 'components/admin/GraphCards/ReactionsByTimeCard/Chart';
+import Chart from 'components/admin/GraphCards/PostsByTimeCard/Chart';
 
 import { isNilOrError } from 'utils/helperUtils';
 
-import NoData from '../../_shared/NoData';
-import messages from '../messages';
-import { TimeSeriesWidgetProps } from '../typings';
+import NoData from '../../../_shared/NoData';
+import messages from '../../messages';
+import { TimeSeriesWidgetProps } from '../../typings';
 
-import useReactionsByTime from './useReactionsByTime';
+import usePostsByTime from './usePostsByTime';
 
-const ReactionsByTime = ({
+const PostsByTime = ({
   projectId,
   startAt,
   endAt,
   resolution = 'month',
 }: TimeSeriesWidgetProps) => {
-  const { currentResolution, timeSeries } = useReactionsByTime({
+  const { currentResolution, timeSeries } = usePostsByTime({
     projectId,
     startAt,
     endAt,
@@ -32,7 +32,7 @@ const ReactionsByTime = ({
 
   return (
     <Box
-      id="e2e-reactions-by-time-widget"
+      className="e2e-posts-by-time-widget"
       width="100%"
       height="220px"
       mt="20px"
@@ -48,4 +48,4 @@ const ReactionsByTime = ({
   );
 };
 
-export default ReactionsByTime;
+export default PostsByTime;
