+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/_deprecated/ReactionsByTimeWidget/ReactionsByTimeCard.tsx
@@ -3,23 +3,23 @@ import React from 'react';
 import { Box } from '@citizenlab/cl2-component-library';
 import moment from 'moment';
 
-import Chart from 'components/admin/GraphCards/CommentsByTimeCard/Chart';
+import Chart from 'components/admin/GraphCards/ReactionsByTimeCard/Chart';
 
 import { isNilOrError } from 'utils/helperUtils';
 
-import NoData from '../../_shared/NoData';
-import messages from '../messages';
-import { TimeSeriesWidgetProps } from '../typings';
+import NoData from '../../../_shared/NoData';
+import messages from '../../messages';
+import { TimeSeriesWidgetProps } from '../../typings';
 
-import useCommentsByTime from './useCommentsByTime';
+import useReactionsByTime from './useReactionsByTime';
 
-const CommentsByTime = ({
+const ReactionsByTime = ({
   projectId,
   startAt,
   endAt,
   resolution = 'month',
 }: TimeSeriesWidgetProps) => {
-  const { currentResolution, timeSeries } = useCommentsByTime({
+  const { currentResolution, timeSeries } = useReactionsByTime({
     projectId,
     startAt,
     endAt,
@@ -32,7 +32,7 @@ const CommentsByTime = ({
 
   return (
     <Box
-      id="e2e-comments-by-time-widget"
+      id="e2e-reactions-by-time-widget"
       width="100%"
       height="220px"
       mt="20px"
@@ -48,4 +48,4 @@ const CommentsByTime = ({
   );
 };
 
-export default CommentsByTime;
+export default ReactionsByTime;
