+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/_deprecated/ReactionsByTimeWidget/index.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import messages from 'containers/Admin/dashboard/messages';
 
-import Card from '../../_shared/Card';
-import TimeSeriesWidgetSettings from '../_shared/TimeSeriesWidgetSettings';
-import { TimeSeriesWidgetProps } from '../typings';
+import Card from '../../../_shared/Card';
+import TimeSeriesWidgetSettings from '../../_shared/TimeSeriesWidgetSettings';
+import { TimeSeriesWidgetProps } from '../../typings';
 
 import ReactionsByTimeCard from './ReactionsByTimeCard';
 
@@ -29,6 +29,8 @@ ReactionsByTimeWidget.craft = {
   },
 };
 
+/** @deprecated This widget should not be used for new reports anymore */
 export const reactionsByTimeTitle = messages.reactions;
 
+/** @deprecated This widget should not be used for new reports anymore */
 export default ReactionsByTimeWidget;
